Move my-account-specific types out of `src/api`
