+++ src/app/my-account/my-account/my-account.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { OffersBaseOffer } from '@app/core/api';
-import { MyAccountPersonalData } from '../../../api/model/my-account-personal-data';
+import { MyAccountPersonalData } from '../my-account.types';
 import { MyAccountService } from '../my-account.service';
 import { Observable } from 'rxjs';
